Pinning/Favoriting Song Implementation
