+++ web/src/pages/api/db/get-user.ts
@@ -1,5 +1,5 @@
 import {NextApiRequest, NextApiResponse} from "next";
-import {authOptions, db} from "../auth/[...nextauth]";
+import {db} from "../auth/[...nextauth]";
 import {RowDataPacket} from "mysql2";
 import {User} from "@/types";
 
@@ -21,14 +21,11 @@ export default async (request: NextApiRequest, response: NextApiResponse) => {
 async function getUsers(user_id: string): Promise<User> {
     try {
         const [userData] = (
-            await db.promise().query(
-                `
+            await db.promise().query(`
             select u.id, u.name, u.email, u.image, u.bio, u.join_date
                 from users u
-                where id=?;
-            `,
-                [user_id]
-            )
+                where id=?;`, 
+            [user_id])
         )[0] as RowDataPacket[];
 
         let user: User = {
